Move helper methods out of the store definitions
