+++ localstack/services/kms/exceptions.py
@@ -1,0 +1,11 @@
+from localstack.aws.api import CommonServiceException
+
+
+class ValidationException(CommonServiceException):
+    def __init__(self, message: str):
+        super().__init__("ValidationException", message, 400, True)
+
+
+class AccessDeniedException(CommonServiceException):
+    def __init__(self, message: str):
+        super().__init__("AccessDeniedException", message, 400, True)
